refactor: update component imports to shared directory
- Changed import paths for various components from the `ui` directory to the `shared` directory, enhancing code organization and consistency.
- Updated imports for Typography, ButtonWithIcon, Icons, and other components across multiple files to reflect this restructuring.
