+++ src/app/(app)/(stacked-layout)/feedback/page.tsx
@@ -1,9 +1,9 @@
 import Image from "next/image";
 
 import { siteConfig } from "@/config/site";
-import { Typography } from "@/components/ui/typography";
 import { SignInToast } from "@/components/feedback/feeback-toast";
 import { FeedbackForm } from "@/components/feedback/feedback-form";
+import { Typography } from "@/components/shared/typography";
 
 export default function FeedbackPage() {
   return (
