+++ src/app/(legal)/privacy-policy/page.mdx
@@ -1,4 +1,4 @@
-import { Typography } from "@/components/ui/typography";
+import { Typography } from "@/components/shared/typography";
 
 # Privacy Policy
