+++ src/components/filters/filter-card/filter-card-meta.tsx
@@ -2,7 +2,7 @@ import type { ConveyorFilterWithAuthor } from "@/types/filter";
 import { cn } from "@/lib/utils";
 import { CardDescription } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
-import { UserBadge } from "@/components/ui/user-badge";
+import { UserBadge } from "@/components/shared/user-badge";
 
 export function FilterCardMeta({
   filter,
