+++ src/components/filters/filter-grid.tsx
@@ -7,9 +7,9 @@ import { useInView } from "react-intersection-observer";
 import type { ConveyorFilterWithAuthor } from "@/types/filter";
 import { useFilters } from "@/hooks/use-filters";
 import { useSearchParams } from "@/hooks/useSearchParams";
-import { Typography } from "@/components/ui/typography";
-import { EmptyState } from "@/components/empty-state";
 import { FilterSortTabs } from "@/components/filters/filter-sort-tabs";
+import { EmptyState } from "@/components/shared/empty-state";
+import { Typography } from "@/components/shared/typography";
 import FiltersLoading from "@/app/(app)/(sidebar-layout)/filters/loading";
 
 import { FilterCard } from "./filter-card/filter-card";
