+++ src/components/footer.tsx
@@ -5,7 +5,7 @@ import { siteConfig } from "@/config/site";
 import { cn } from "@/lib/utils";
 import { Button, buttonVariants } from "@/components/ui/button";
 
-import { Icons } from "./ui/icons";
+import { Icons } from "./shared/icons";
 
 const navigation = [
   { name: "Privacy Policy", href: "/privacy-policy" },
