+++ src/components/header.tsx
@@ -11,13 +11,12 @@ import {
 import { Loader2 } from "lucide-react";
 
 import { siteConfig } from "@/config/site";
-import { RepoStarsButton } from "@/components/repo-stars-button";
-
-import { DesktopNav } from "./desktop-nav";
-import { MobileNav } from "./mobile-nav";
-import { ModeToggle } from "./mode-toggle";
-import { Button } from "./ui/button";
-import { Icons } from "./ui/icons";
+import { Button } from "@/components/ui/button";
+import { DesktopNav } from "@/components/desktop-nav";
+import { MobileNav } from "@/components/mobile-nav";
+import { Icons } from "@/components/shared/icons";
+import { ModeToggle } from "@/components/shared/mode-toggle";
+import { RepoStarsButton } from "@/components/shared/repo-stars-button";
 
 export type Navigation = {
   name: string;
