+++ src/components/shared/banner.tsx
