+++ src/components/shared/text-effect.tsx
