+++ src/components/shared/theme-image.tsx
