+++ src/components/shared/typography.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import { Slot as SlotPrimitive } from "radix-ui";
 import { cva, VariantProps } from "class-variance-authority";
+import { Slot as SlotPrimitive } from "radix-ui";
 
 import { cn } from "@/lib/utils";
